+++ src/bot/botrulez/ping.rs
@@ -30,11 +30,13 @@ where
         msg: &Message,
         ctx: &Context,
         _bot: &mut B,
-    ) -> Result<(), E> {
+    ) -> Result<bool, E> {
         if arg.trim().is_empty() {
             ctx.reply(msg.id, &self.0).await?;
+            Ok(true)
+        } else {
+            Ok(false)
         }
-        Ok(())
     }
 }
 
@@ -55,8 +57,8 @@ where
         msg: &Message,
         ctx: &Context,
         _bot: &mut B,
-    ) -> Result<(), E> {
+    ) -> Result<bool, E> {
         ctx.reply(msg.id, &self.0).await?;
-        Ok(())
+        Ok(true)
     }
 }